Ajout handlebar

+++ src/main/java/Build.java
@@ -4,20 +4,29 @@ Groupe : PRSV
 Description : Implémentation cmd Build
  */
 
+import com.github.jknack.handlebars.Handlebars;
+import com.github.jknack.handlebars.Template;
+import com.github.jknack.handlebars.io.FileTemplateLoader;
+import com.github.jknack.handlebars.io.TemplateLoader;
 import org.tautua.markdownpapers.Markdown;
 import org.tautua.markdownpapers.parser.ParseException;
 import picocli.CommandLine;
 
 import java.io.*;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.concurrent.Callable;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import org.apache.commons.io.FileUtils;
+import utils.HandlebarUtil;
+import utils.JSONConfig;
 
 @CommandLine.Command(
         name = "build",
@@ -52,13 +61,18 @@ public class Build implements Callable<Integer> {
 
         //Copie intégrale des fichiers présent dans build, puis conversion et clean du dossier
         File folder = new File(rootDirectory + "/build/");
-        File contentFile = new File (folder+"/template");
+        File contentFile = new File (folder + "/" + Constantes.TEMPLATE_DIRECTORY);
         if(contentFile.exists())
         {
             FileUtils.forceDelete(contentFile); //Suppression du dossier template
         }
 
-        try(Stream<Path> walk = Files.walk(Paths.get(rootDirectory+"/build/"))){
+        //Lecture fichier de configuration
+        JSONConfig config = new JSONConfig(rootDirectory + "/" + Constantes.CONFIG_FILE_NAME);
+        config.read();
+        String configTitre = config.getTitre();
+
+        try(Stream<Path> walk = Files.walk(Paths.get(rootDirectory + "/build/"))){
             List<File> result = walk.filter(Files::isRegularFile)
                     .map(x -> x.toFile()).collect(Collectors.toList());
             for(File f : result)
@@ -69,21 +83,53 @@ public class Build implements Callable<Integer> {
                     String extension = filename.substring(index + 1);
                     if(extension.equals("md")) //Si le fichier est sous format md, il est convertit en html
                     {
-                        Reader in = new FileReader(f);
+                        BufferedReader br = new BufferedReader(new InputStreamReader(new FileInputStream(f), StandardCharsets.UTF_8));
+                        String line;
+                        line = br.readLine();
+                        if(line.startsWith("titre : ")){
 
-                        filename = f.getPath();
-                        filename = filename.replace("md","html");
+                            //INline
+                            String titrePage = line.substring(8);// process the line.
+                            Handlebars handlebars = new Handlebars();
+                            Template template = handlebars.compileInline("Hi {{site}}! Hi {{page}}! Hi {{content}}!");
 
-                        Writer out = new FileWriter(filename);
-                        Markdown md = new Markdown();
-                        md.transform(in, out);
+                            //Création de la map
+                            Map<String, String> parameterMap = new HashMap<>();
 
-                        out.close();
-                        f.delete();
+                            //Pair clé valeur name => Baeldung
+                            parameterMap.put("site", configTitre);
+                            parameterMap.put("page", titrePage);
+                            parameterMap.put("content", "contenu");
+
+                            HandlebarUtil handlebarLocal = new  HandlebarUtil(rootDirectory);
+                            handlebarLocal.transform(parameterMap);
+
+                            //On injectle contenu de la map sur le template
+                            String templateString = template.apply(parameterMap);
+                            System.out.println(templateString);
+
+
+                            Reader in = new FileReader(f);
+
+                            filename = f.getPath();
+                            filename = filename.replace("md","html");
+
+                            Writer out = new FileWriter(filename);
+                            Markdown md = new Markdown();
+                            md.transform(in, out);
+
+                            out.close();
+                            f.delete();
+
+
+
+                        }else{
+                            throw new IllegalArgumentException("La 1ère ligne doit être le titre");
+                        }
                     }
                     else if(f.getName().equals(Constantes.CONFIG_FILE_NAME)) //On elimine le fichier de config
                     {
-                        f.delete(); //
+                        f.delete();
                     }
                 }
             }
